+++ src/main/java/dev/doctor4t/ratatouille/client/render/feature/RootCustomBipedArmorFeatureRenderer.java
@@ -6,11 +6,13 @@ import dev.doctor4t.ratatouille.client.util.CustomModelArmorUtil;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
 import net.minecraft.client.render.OverlayTexture;
 import net.minecraft.client.render.RenderLayer;
+import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.render.VertexConsumerProvider;
 import net.minecraft.client.render.entity.EntityRendererFactory;
 import net.minecraft.client.render.entity.feature.FeatureRenderer;
 import net.minecraft.client.render.entity.feature.FeatureRendererContext;
 import net.minecraft.client.render.entity.model.BipedEntityModel;
+import net.minecraft.client.render.item.ItemRenderer;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.util.Identifier;
@@ -33,34 +35,20 @@ public class RootCustomBipedArmorFeatureRenderer<T extends LivingEntity, M exten
 
     @Override
     public void render(MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, T entity, float limbAngle, float limbDistance, float tickDelta, float animationProgress, float headYaw, float headPitch) {
-        for (ArmorDisplayConditions customArmorSetItems : customArmorModels.keySet()) {
-            CustomBipedArmorModel<T> customArmorModel = (CustomBipedArmorModel<T>) customArmorModels.get(customArmorSetItems);
+        for (ArmorDisplayConditions customArmorSetConditions : customArmorModels.keySet()) {
+            CustomBipedArmorModel<T> customArmorModel = (CustomBipedArmorModel<T>) customArmorModels.get(customArmorSetConditions);
 
-            // rotations
             this.getContextModel().copyBipedStateTo(customArmorModel);
 
-            // check visibility
-            boolean shouldDisplayHelmet = customArmorSetItems.shouldDisplayHelmet(entity);
-            boolean shouldDisplayChestplate = customArmorSetItems.shouldDisplayChestplate(entity);
-            boolean shouldDisplayLeggings = customArmorSetItems.shouldDisplayLeggings(entity);
-            boolean shouldDisplayBoots = customArmorSetItems.shouldDisplayBoots(entity);
+            VertexConsumer helmetBuffer = ItemRenderer.getArmorGlintConsumer(vertexConsumers, RenderLayer.getEntityTranslucent(customArmorModel.getTexture()), customArmorSetConditions.shouldDisplayHelmetGlint(entity));
+            VertexConsumer chestplateBuffer = ItemRenderer.getArmorGlintConsumer(vertexConsumers, RenderLayer.getEntityTranslucent(customArmorModel.getTexture()), customArmorSetConditions.shouldDisplayChestplateGlint(entity));
+            VertexConsumer leggingsBuffer = ItemRenderer.getArmorGlintConsumer(vertexConsumers, RenderLayer.getEntityTranslucent(customArmorModel.getTexture()), customArmorSetConditions.shouldDisplayLeggingsGlint(entity));
+            VertexConsumer bootsBuffer = ItemRenderer.getArmorGlintConsumer(vertexConsumers, RenderLayer.getEntityTranslucent(customArmorModel.getTexture()), customArmorSetConditions.shouldDisplayBootsGlint(entity));
 
-            // apply visibility
-            customArmorModel.helmet.visible = shouldDisplayHelmet;
-
-            customArmorModel.body_chestplate.visible = shouldDisplayChestplate;
-            customArmorModel.right_arm_chestplate.visible = shouldDisplayChestplate;
-            customArmorModel.left_arm_chestplate.visible = shouldDisplayChestplate;
-
-            customArmorModel.body_leggings.visible = shouldDisplayLeggings;
-            customArmorModel.right_leg_leggings.visible = shouldDisplayLeggings;
-            customArmorModel.left_leg_leggings.visible = shouldDisplayLeggings;
-
-            customArmorModel.right_leg_boot.visible = shouldDisplayBoots;
-            customArmorModel.left_leg_boot.visible = shouldDisplayBoots;
-
-            // render
-            customArmorModel.render(matrices, vertexConsumers.getBuffer(RenderLayer.getEntityTranslucent(customArmorModel.getTexture())), light, OverlayTexture.DEFAULT_UV);
+            if (customArmorSetConditions.shouldDisplayHelmet(entity)) customArmorModel.renderHelmet(matrices, helmetBuffer, light, OverlayTexture.DEFAULT_UV);
+            if (customArmorSetConditions.shouldDisplayChestplate(entity)) customArmorModel.renderChestplate(matrices, chestplateBuffer, light, OverlayTexture.DEFAULT_UV);
+            if (customArmorSetConditions.shouldDisplayLeggings(entity)) customArmorModel.renderLeggings(matrices, leggingsBuffer, light, OverlayTexture.DEFAULT_UV);
+            if (customArmorSetConditions.shouldDisplayBoots(entity)) customArmorModel.renderBoots(matrices, bootsBuffer, light, OverlayTexture.DEFAULT_UV);
         }
     }