+++ src/main/java/dev/doctor4t/ratatouille/client/util/ItemSetDisplayConditions.java
@@ -21,15 +21,36 @@ public class ItemSetDisplayConditions extends ArmorDisplayConditions {
         return livingEntity.getEquippedStack(EquipmentSlot.HEAD).isOf(this.helmetItem);
     }
 
+    @Override
+    public boolean shouldDisplayHelmetGlint(LivingEntity livingEntity) {
+        return livingEntity.getEquippedStack(EquipmentSlot.HEAD).hasGlint();
+    }
+
     public boolean shouldDisplayChestplate(LivingEntity livingEntity) {
         return livingEntity.getEquippedStack(EquipmentSlot.CHEST).isOf(this.chesplateItem);
+    }
+
+    @Override
+    public boolean shouldDisplayChestplateGlint(LivingEntity livingEntity) {
+        return livingEntity.getEquippedStack(EquipmentSlot.CHEST).hasGlint();
     }
 
     public boolean shouldDisplayLeggings(LivingEntity livingEntity) {
         return livingEntity.getEquippedStack(EquipmentSlot.LEGS).isOf(this.leggingsItem);
     }
 
+    @Override
+    public boolean shouldDisplayLeggingsGlint(LivingEntity livingEntity) {
+        return livingEntity.getEquippedStack(EquipmentSlot.LEGS).hasGlint();
+    }
+
     public boolean shouldDisplayBoots(LivingEntity livingEntity) {
         return livingEntity.getEquippedStack(EquipmentSlot.FEET).isOf(this.bootsItem);
     }
+
+    @Override
+    public boolean shouldDisplayBootsGlint(LivingEntity livingEntity) {
+        return livingEntity.getEquippedStack(EquipmentSlot.FEET).hasGlint();
+    }
+
 }